# openapi.yml
openapi: 3.1.0
info:
  title: AgoraLume API
  description: |-
    User + multi-AI group chat with modular personas. The backend owns the workspace and streams chat over SSE.

    POST creates (201), PATCH merges a partial body (200), DELETE returns 204. Failures are RFC 9457 problem documents; `type` is a stable `urn:agoralume:error:…`.
  license:
    name: AGPL-3.0-only
    identifier: AGPL-3.0-only
  version: 0.1.0
servers:
- url: /
  description: 'Same-origin: the bundled build, where the SPA and the API share a host'
- url: /api
  description: Behind an edge/dev proxy (VITE_API_PREFIX=/api)
- url: http://127.0.0.1:8080
  description: A local backend (pnpm dev:api)
paths:
  /v1beta/accounts:
    get:
      tags:
      - accounts
      summary: Every existing account, for the admin dashboard's account list.
      operationId: list_accounts
      responses:
        '200':
          description: Every account's id and login name; never a password
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AccountSummary'
        '401':
          description: Not signed in
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: Not permitted for this identity
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiError'
    post:
      tags:
      - accounts
      summary: |-
        Provisions a brand-new account with an admin-chosen username and
        password. There's no self-service registration — this is the only way an
        account gets created.
      operationId: create_account
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAccountRequest'
        required: true
      responses:
        '201':
          description: The created account
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountSummary'
        '401':
          description: Not signed in
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: Not permitted for this identity
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '422':
          description: Empty/reserved/taken username, empty password, or no persistent backend
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /v1beta/accounts/{accountId}:
    patch:
      tags:
      - accounts
      summary: |-
        Changes an existing account's username, password, or both. Every live
        session for that account is revoked: its tokens stop working immediately and
        the holder must sign in again.
      operationId: update_account
      parameters:
      - name: accountId
        in: path
        description: The account to edit
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAccountRequest'
        required: true
      responses:
        '200':
          description: The updated account; its sessions are revoked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountSummary'
        '401':
          description: Not signed in
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: Not permitted for this identity
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '422':
          description: Unknown id, empty/reserved/taken username, empty password, or no persistent backend
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiError'
...
  /v1beta/auth/refresh:
    post:
      tags:
      - auth
      summary: |-
        Mints a fresh access token from a still-valid refresh token, without
        asking for the password again. The refresh token itself is not rotated —
        it keeps working until its own (much longer) expiry.
      operationId: refresh
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
        required: true
      responses:
        '200':
          description: A fresh access token; the refresh token is unchanged
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccessToken'
        '401':
          description: Unknown, expired, revoked, or not a refresh token
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiError'
      security:
      - {}
...
  /v1beta/groups/{groupId}/messages:
    get:
      tags:
      - chat
      summary: |-
        A contiguous window of message history, oldest first. Max 500 lines (160
        without an `anchor`).
      operationId: list_messages
      parameters:
      - name: groupId
        in: path
        description: The group whose history to read
        required: true
        schema:
          type: string
      - name: anchor
        in: query
        description: Line id to centre the window on. Omitted, the window ends at the newest line.
        required: false
        schema:
          type:
          - string
          - 'null'
      - name: before
        in: query
        description: Lines before the anchor (or before the tail). Clamped to 500.
        required: false
        schema:
          type:
          - integer
          - 'null'
          maximum: 500
          minimum: 0
      - name: after
        in: query
        description: Lines after the anchor. Ignored without an `anchor`. Clamped to 500.
        required: false
        schema:
          type:
          - integer
          - 'null'
          maximum: 500
          minimum: 0
      - name: since
        in: query
        description: |-
          Read mark (epoch ms). Without an `anchor`, widens the window back to cover
          every unread line.
        required: false
        schema:
          type:
          - integer
          - 'null'
          format: int64
      responses:
        '200':
          description: The window, oldest first; empty for an unknown group
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Message'
        '401':
          description: Not signed in
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: Not permitted for this identity
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiError'
    post:
      tags:
      - chat
      summary: |-
        Posts a user message and starts the agents' turn. Replies, moods, and read
        receipts arrive on the SSE stream.
      operationId: send_message
      parameters:
      - name: groupId
        in: path
        description: The group to post into
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendBody'
        required: true
      responses:
        '200':
          description: The stored message
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '401':
          description: Not signed in
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: Not permitted for this identity
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: Unknown group
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '422':
          description: Blank or over-long text, or a `personaId` that is not a user identity
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiError'
...
components:
  schemas:
    AccessToken:
      type: object
      required:
      - accessToken
      properties:
        accessToken:
          type: string
    AccountSummary:
      type: object
      description: |-
        One account, as admin sees it: its opaque id and its login name. Never
        carries a password or its hash. `account_id` is what [`UpdateAccountRequest`]
        targets via `PATCH /accounts/{account_id}`.
      required:
      - accountId
      - username
      properties:
        accountId:
          type: string
        username:
          type: string
...
    ApiError:
      type: object
      description: |-
        An RFC 9457 problem document — the body of every 4xx this API produces.

        Constructed through the named helpers below rather than field-by-field, so
        the slug used for a given status is decided in exactly one place and can't
        drift between handlers.
      required:
      - type
      - title
      - status
      properties:
        detail:
          type:
          - string
          - 'null'
          description: |-
            The human-readable explanation of *this specific* occurrence — the
            sentence that used to be the entire `text/plain` body. Safe to show a
            user; never carries a secret, a stack trace, or a provider's raw
            response.
        status:
          type: integer
          format: int32
          description: |-
            The HTTP status code, repeated in the body so a problem document stays
            meaningful when it's logged or forwarded away from its response.
          minimum: 0
        title:
          type: string
          description: |-
            The status code's canonical reason phrase (e.g. `Forbidden`) — a short
            human summary of the problem *type*, not of this occurrence.
        type:
          type: string
          description: |-
            Stable, machine-readable identifier for *what* went wrong, e.g.
            `urn:agoralume:error:not-admin`. Match on this rather than on `detail`,
            whose wording is free to change. Always present.
          example: urn:agoralume:error:not-found
...
    CreateAccountRequest:
      type: object
      description: |-
        The `POST /accounts` request: the initial login for a brand-new account,
        chosen by admin and communicated to whoever the account is for out of
        band — there's no self-service registration or password-reset flow.
      required:
      - username
      - password
      properties:
        password:
          type: string
        username:
          type: string
...
    Message:
      oneOf:
      - type: object
        description: A normal, readable message.
        required:
        - id
        - groupId
        - personaId
        - ts
        - text
        - kind
        properties:
          groupId:
            type: string
          id:
            type: string
          kind:
            type: string
            enum:
            - conversation
          personaId:
            type: string
          readBy:
            type:
            - array
            - 'null'
            items:
              type: string
            description: AI persona ids that have successfully processed (read) this message.
          streaming:
            type:
            - boolean
            - 'null'
          text:
            type: string
          ts:
            type: integer
            format: int64
      - type: object
        description: A persona broadcasting its current mood/emotion.
        required:
        - id
        - groupId
        - personaId
        - ts
        - mood
        - kind
        properties:
          groupId:
            type: string
          id:
            type: string
          kind:
            type: string
            enum:
            - mood
          mood:
            type: string
          note:
            type:
            - string
            - 'null'
          personaId:
            type: string
          ts:
            type: integer
            format: int64
      - type: object
        description: |-
          A system notice that an agent's inference failed after exhausting retries.
          Surfaces the HTTP status and canonical reason only — never the provider's
          raw body, which can leak quota or key details. `personaId` is the agent
          that failed. The frontend renders it as an error line with a retry button.
        required:
        - id
        - groupId
        - personaId
        - ts
        - reason
        - kind
        properties:
          groupId:
            type: string
          id:
            type: string
          kind:
            type: string
            enum:
            - system
          personaId:
            type: string
          reason:
            type: string
            description: |-
              The status's canonical reason (e.g. "Too Many Requests"), or a short
              generic label for a transport failure with no status.
          status:
            type:
            - integer
            - 'null'
            format: int32
            description: The HTTP status code, when the failure carried one (e.g. 429).
            minimum: 0
          ts:
            type: integer
            format: int64
      description: One chat line. Tagged by `kind`, matching the TypeScript `Message` union.
...
    RefreshRequest:
      type: object
      required:
      - refreshToken
      properties:
        refreshToken:
          type: string
    SendBody:
      type: object
      description: The body of a send request.
      required:
      - text
      properties:
        personaId:
          type:
          - string
          - 'null'
          description: |-
            Which user identity to author as; must be a `user` persona, not an AI
            one. Omitted, the group's `selfPersonaId` is used.
        text:
          type: string
          description: The message text.
          maxLength: 8000
...
    UpdateAccountRequest:
      type: object
      description: |-
        The `PATCH /accounts/{account_id}` request: admin changing an existing
        account's login — its username, its password, or both. Both fields are
        optional and independent (`None` leaves that field untouched); an empty
        patch is a harmless no-op. There's still no self-service path — this is
        admin editing *someone else's* login, not an account changing its own.
      properties:
        password:
          type:
          - string
          - 'null'
        username:
          type:
          - string
          - 'null'
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: 'Access token from `POST /auth/login`. Opaque; does not survive a server restart. Waived when `GET /meta` reports `authRequired: false`.'
security:
- bearerAuth: []
...